requirements.txt
pandas
numpy
python-dotenv
yfinance
requests
transformers
torch
python-dotenv
